Bonus Objective 1 DONE
I stored an environment variable in a env. file and used it within the shopping cart code. Now the end user can input their specific tax rate and use it in the program. I also updated the README.md file to instruct the user how to use the variable. I added an email text page to start working on the second bonus objective.